requirements.txt
Flask
Flask-SQLAlchemy
pymysql
langchain-google-genai
google-generativeai
pydantic
spacy
groq
langchain-groq
langchain
langchain-huggingface
langchain-community
pinecone-client
sentence-transformers
tf-keras
tiktoken
yfinance
mysqlclient
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz